refactor: add explicit types
