+++ src/hooks/useBalanceChartData/useBalanceChartData.ts
@@ -23,7 +23,7 @@ import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { priceAtDate } from 'lib/charts'
 import { logger } from 'lib/logger'
 import type { AssetsById } from 'state/slices/assetsSlice/assetsSlice'
-import type { PriceHistoryData } from 'state/slices/marketDataSlice/marketDataSlice'
+import type { PriceHistoryData } from 'state/slices/marketDataSlice/types'
 import type { AssetBalancesById } from 'state/slices/portfolioSlice/portfolioSliceCommon'
 import {
   selectAssets,
